moved sprite method to image module
- trying other fonts
