+++ examples/spritesheet1/index.js
@@ -12,7 +12,8 @@ const spritesheet1 = {
         //await points.addTextureImage('image', './../img/old_king_600x600.jpg', ShaderType.FRAGMENT);
         //await points.addTextureImage('image', './../assets_ignore/absulit_800x800.jpg', ShaderType.FRAGMENT);
         // await points.addTextureImage('image', './../img/sprite_nums_640x640.png', ShaderType.FRAGMENT);
-        await points.addTextureImage('image', './../assets_ignore/pixelspritefont 32_green.png', ShaderType.FRAGMENT);
+        // await points.addTextureImage('image', './../assets_ignore/pixelspritefont 32_green.png', ShaderType.FRAGMENT);
+        await points.addTextureImage('image', './../assets_ignore/pixeldroid_botic_64x64.png', ShaderType.FRAGMENT);
     },
     update: (points) => {
